extract fixtures module, create testing feature
