+++ muxink/Cargo.toml
@@ -3,6 +3,17 @@ name = "muxink"
 version = "0.1.0"
 edition = "2021"
 
+[features]
+default = []
+testing = ["tokio-stream"]
+
+[[bin]]
+name = "load_testing"
+src = "bin/load_testing.rs"
+test = false
+bench = false
+required-features = ["testing"]
+
 [dependencies]
 bytes = "1.1.0"
 futures = "0.3.21"
@@ -10,6 +21,7 @@ thiserror = "1.0.31"
 tokio = { version = "1" }
 tokio-util = "0.7.2"
 tracing = "0.1.18"
+tokio-stream = { version = "0.1.8", optional = true }
 
 [dev-dependencies]
 tokio = { version = "1", features = [ "io-util", "macros", "net", "rt" ] }
