+++ muxink/src/bin/load_testing.rs
@@ -1,0 +1,5 @@
+use muxink;
+
+fn main() {
+    println!("hello world");
+}
